MVP-4710-patch-6: create new hook for web3Target
