+++ packages/notifi-react/lib/context/NotifiTargetContext.tsx
@@ -13,9 +13,7 @@ import React, {
   useState,
 } from 'react';
 
-import { defaultCopy } from '../utils';
-import { reformatSignatureForWalletTarget } from '../utils/wallet';
-import { createCoinbaseNonce, subscribeCoinbaseMessaging } from '../utils/xmtp';
+import { useTargetWallet } from '../hooks/useTargetWallet';
 import { useNotifiFrontendClientContext } from './NotifiFrontendClientContext';
 
 export type TargetGroupInput = {
@@ -145,8 +143,10 @@ const isToggleTargetRenewArgs = (
 };
 
 export type NotifiTargetContextType = {
-  isLoading: boolean;
+  isLoading: boolean; // general loading state: updateTargetDocument, initial load, renewTargetGroup
+  isLoadingWallet: boolean; // wallet target loading state: verify wallet target
   error: Error | null;
+  errorWallet: Error | null;
   updateTargetInputs: UpdateTargetInputs;
   renewTargetGroup: (singleTargetRenewArgs?: {
     target: ToggleTarget;
@@ -169,9 +169,8 @@ export type NotifiTargetContextProviderProps = {
 export const NotifiTargetContextProvider: FC<
   PropsWithChildren<NotifiTargetContextProviderProps>
 > = ({ children, toggleTargetAvailability }) => {
-  const { frontendClient, frontendClientStatus, walletWithSignParams } =
+  const { frontendClient, frontendClientStatus } =
     useNotifiFrontendClientContext();
-  const xmtp = useClient();
 
   const [error, setError] = useState<Error | null>(null);
   const [isLoading, setIsLoading] = useState(true);
@@ -212,6 +211,11 @@ export const NotifiTargetContextProvider: FC<
       isAvailable: toggleTargetAvailability?.wallet ?? false,
     },
   });
+  const {
+    signCoinbaseSignature,
+    isLoading: isLoadingWallet,
+    error: errorWallet,
+  } = useTargetWallet(targetData.wallet);
   const [targetInfoPrompts, setTargetInfoPrompts] = useState<
     Partial<Record<Target, TargetInfo>>
   >({
@@ -712,201 +716,6 @@ export const NotifiTargetContextProvider: FC<
     [],
   );
 
-  const senderAddress =
-    // TODO: refactor this const into needed methods
-    // eslint-disable-next-line
-    // @ts-ignore TODO: diable after BE supports senderAddress
-    targetData?.wallet?.data?.senderAddress ??
-    defaultCopy.notifiTargetContext.defaultWalletTargetSenderAddress; // Default with GMX address
-
-  const getSignature = useCallback(
-    async (message: Uint8Array | string) => {
-      let signature: Uint8Array | string = '';
-
-      if (typeof message === 'string') {
-        const encoder = new TextEncoder();
-        message = encoder.encode(message);
-      }
-
-      // TODO: Add logic for rest of the chains
-      switch (walletWithSignParams.walletBlockchain) {
-        case 'AVALANCHE':
-        case 'ETHEREUM':
-        case 'POLYGON':
-        case 'ARBITRUM':
-        case 'BINANCE':
-        case 'ELYS':
-        case 'NEUTRON':
-        case 'ARCHWAY':
-        case 'AXELAR':
-        case 'BERACHAIN':
-        case 'OPTIMISM':
-        case 'ZKSYNC':
-        case 'INJECTIVE':
-        case 'BASE':
-        case 'BLAST':
-        case 'CELO':
-        case 'MANTLE':
-        case 'LINEA':
-        case 'SCROLL':
-        case 'MANTA':
-        case 'EVMOS':
-        case 'MONAD':
-        case 'AGORIC':
-        case 'ORAI':
-        case 'KAVA':
-        case 'CELESTIA':
-        case 'COSMOS':
-        case 'DYMENSION':
-        case 'DYDX':
-        case 'XION':
-        case 'NEAR':
-        case 'SUI':
-          signature = await walletWithSignParams.signMessage(message);
-          break;
-        default: {
-          setError(Error('This chain is not supported'));
-          throw Error('This chain is not supported');
-        }
-      }
-      return reformatSignatureForWalletTarget(signature);
-    },
-    [walletWithSignParams.signMessage],
-  );
-
-  // NOTE: Temporarily commenting out this function because it will be needed later
-  // const xip43Impl = async () => {
-
-  //   const targetId = targetData?.wallet?.data?.id ?? '';
-  //   const address = '';
-
-  //   const timestamp = Date.now();
-  //   const message = createConsentMessage(senderAddress, timestamp);
-  //   const signature = '';
-
-  //   if (!signature) {
-  //     throw Error('Unable to sign the wallet. Please try again.');
-  //   }
-
-  //   await frontendClient.verifyXmtpTarget({
-  //     input: {
-  //       web3TargetId: targetId,
-  //       accountId: address,
-  //       consentProofSignature: signature as string,
-  //       timestamp: timestamp,
-  //       isCBW: true,
-  //     },
-  //   });
-  //   // await signCoinbaseSignature(address, senderAddress);
-  //   await frontendClient.verifyCbwTarget({
-  //     input: {
-  //       targetId: targetId,
-  //     },
-  //   });
-  // };
-
-  const xmtpXip42Impl = useCallback(async () => {
-    type XmtpInitOption = (typeof xmtp.initialize extends (a: infer U) => void
-      ? U
-      : never)['options'];
-
-    const options: XmtpInitOption = {
-      persistConversations: false,
-      env: 'production',
-    };
-    const address = walletWithSignParams.walletPublicKey;
-
-    const signer = {
-      getAddress: (): Promise<string> => {
-        return new Promise((resolve) => {
-          resolve(address);
-        });
-      },
-      signMessage: async (message: Uint8Array | string): Promise<string> => {
-        return getSignature(message);
-      },
-    };
-    // NOTE: 1st signature: init XMTP with user wallet (need sign every time)
-    const client = await xmtp.initialize({ options, signer });
-
-    if (client === undefined) {
-      throw Error('XMTP client is uninitialized. Please try again.');
-    }
-    // NOTE: 2nd signature: create a new XMTP conversation with the tenant sender (will skip if ever signed before)
-    const conversation = await client.conversations.newConversation(
-      senderAddress,
-    );
-
-    await client.contacts.allow([senderAddress]);
-
-    return conversation.topic.split('/')[3];
-  }, [walletWithSignParams.walletPublicKey, xmtp, getSignature]);
-
-  const signCoinbaseSignature = useCallback(
-    async (web3TargetId: Types.Web3Target['id']) => {
-      setIsLoading(true);
-      try {
-        const conversationTopic = await xmtpXip42Impl();
-
-        const address = walletWithSignParams.walletPublicKey;
-        let nonce = '';
-        try {
-          nonce = await createCoinbaseNonce();
-        } catch (e) {
-          console.error('error in createCoinbaseNonce: ', e);
-          if (e instanceof Error) return setError(e);
-        }
-        if (!nonce)
-          return setError(
-            Error('Unable to get the nonce from coinbase wallet, try again.'),
-          );
-
-        const message = `Coinbase Wallet Messaging subscribe\nAddress: ${address}\nPartner Address: ${senderAddress}\nNonce: ${nonce}`;
-
-        // NOTE: 3rd signature: sign the notifi message above to sync with Notifi BE
-        const signature = await getSignature(message);
-
-        if (!signature)
-          throw Error('Unable to sign the wallet. Please try again.');
-
-        const payload = {
-          address,
-          nonce,
-          signature: signature as `0x${string}`,
-          isActivatedViaCb: true,
-          partnerAddress: senderAddress,
-          conversationTopic,
-        };
-        await subscribeCoinbaseMessaging(payload);
-        await frontendClient.verifyXmtpTargetViaXip42({
-          input: {
-            web3TargetId: web3TargetId,
-            accountId: address,
-            conversationTopic,
-          },
-        });
-
-        await frontendClient
-          .fetchData()
-          .then(refreshTargetDocument)
-          .catch((e: unknown) => console.error(e));
-
-        return true;
-      } catch (e) {
-        console.error('error in signCoinbaseSignature: ', e);
-        return false;
-      } finally {
-        setIsLoading(false);
-      }
-    },
-    [
-      walletWithSignParams.walletPublicKey,
-      frontendClient,
-      xmtpXip42Impl,
-      getSignature,
-    ],
-  );
-
   const refreshWeb3Target = useCallback(
     async (web3Target?: Types.Web3TargetFragmentFragment) => {
       if (web3Target) {
@@ -915,9 +724,13 @@ export const NotifiTargetContextProvider: FC<
             updateTargetInfoPrompt('wallet', {
               type: 'cta',
               message: 'Sign Wallet',
-              onClick: () => {
+              onClick: async () => {
                 // NOTE: sign coinbase requires up to 3 signing process: 1. init XMTP, 2. create XMTP conversation, 3. sign the confirm message to Notifi BE
-                return signCoinbaseSignature(web3Target.id);
+                const updatedWeb3Target = await signCoinbaseSignature(
+                  web3Target.id,
+                );
+                if (!updatedWeb3Target) return;
+                refreshWeb3Target(updatedWeb3Target);
               },
             });
             break;
@@ -977,7 +790,9 @@ export const NotifiTargetContextProvider: FC<
       value={{
         refreshTargetDocument,
         error,
+        errorWallet,
         isLoading,
+        isLoadingWallet,
         renewTargetGroup,
         unVerifiedTargets,
         isChangingTargets,
